+++ examples/app.rb
@@ -7,12 +7,12 @@ require 'rack'
 
 App = Rack::Builder.new do
   use OpenapiFirst::RequestValidation, raise_error: true, spec: File.expand_path('./openapi.yaml', __dir__)
-  use OpenapiFirst::ResponseValidation
+  use OpenapiFirst::ResponseValidation, spec: File.expand_path('./openapi.yaml', __dir__)
 
   handlers = {
     'things#index' => ->(_env) { [200, { 'Content-Type' => 'application/json' }, ['{"hello": "world"}']] }
   }
   not_found = ->(_env) { [404, {}, []] }
 
-  run ->(env) { handlers.fetch(env[OpenapiFirst::OPERATION].operation_id, not_found).call(env) }
+  run ->(env) { handlers.fetch(env[OpenapiFirst::REQUEST].operation_id, not_found).call(env) }
 end