+++ lib/openapi_first/request_validation/failure.rb
@@ -4,6 +4,8 @@ module OpenapiFirst
   module RequestValidation
     class Failure
       def initialize(status:, location:, message: nil, validation_result: nil)
+        raise ArgumentError, ":location must be one of #{TOPICS.keys}" unless TOPICS.key?(location)
+
         @status = status
         @location = location
         @message = message
@@ -16,11 +18,20 @@ module OpenapiFirst
         @message || validation_result&.message || Rack::Utils::HTTP_STATUS_CODES[status]
       end
 
+      def raise!
+        raise NotFoundError, error_message if location == :not_found
+
+        raise RequestInvalidError, error_message
+      end
+
+      private
+
       def error_message
         "#{TOPICS.fetch(location)} #{message}"
       end
 
       TOPICS = {
+        not_found: 'Request not defined.',
         body: 'Request body invalid:',
         query: 'Query parameter invalid:',
         header: 'Header parameter invalid:',