+++ lib/openapi_first/request_validation/request_body_validator.rb
@@ -1,33 +1,32 @@
 # frozen_string_literal: true
 
-require_relative 'body_parser'
-
 module OpenapiFirst
   module RequestValidation
     class RequestBodyValidator
-      def initialize(operation, env)
+      def initialize(operation)
         @operation = operation
-        @env = env
       end
 
-      def validate!
-        request_body = @operation.request_body
+      def validate!(runtime_request)
+        request_body = operation.request_body
         return unless request_body
 
-        request_content_type = Rack::Request.new(@env).content_type
+        request_content_type = runtime_request.content_type
         schema = request_body.schema_for(request_content_type)
         RequestValidation.fail!(:header, status: 415) unless schema
 
-        parsed_request_body = BodyParser.new.parse_body(@env)
+        parsed_request_body = runtime_request.body
         RequestValidation.fail!(:body) if request_body.required? && parsed_request_body.nil?
 
         validate_body!(parsed_request_body, schema)
         parsed_request_body
-      rescue BodyParsingError => e
+      rescue BodyParser::ParsingError => e
         RequestValidation.fail!(:body, message: e.message)
       end
 
       private
+
+      attr_reader :operation
 
       def validate_body!(parsed_request_body, schema)
         request_body_schema = schema