+++ spec/request_validation/header_parameter_validation_spec.rb
@@ -43,7 +43,7 @@ RSpec.describe 'Header Parameter validation' do
     it 'adds the converted header parameter to env ' do
       header 'Accept-Version', '1'
       get '/pets'
-      expect(last_request.env[OpenapiFirst::HEADER_PARAMS]['Accept-Version']).to eq 1
+      expect(last_request.env[OpenapiFirst::REQUEST].headers['Accept-Version']).to eq 1
     end
 
     describe 'when raising' do