+++ spec/request_validation/path_parameter_validation_spec.rb
@@ -33,7 +33,7 @@ RSpec.describe 'Path Parameter validation' do
 
     it 'adds the converted path parameter to env ' do
       get '/pets/42'
-      expect(last_request.env[OpenapiFirst::PARAMS]['petId']).to eq 42
+      expect(last_request.env[OpenapiFirst::REQUEST].params['petId']).to eq 42
     end
 
     it 'succeeds if path parameter are valid' do