+++ spec/request_validation/query_parameter_validation_spec.rb
@@ -95,7 +95,7 @@ RSpec.describe 'Query Parameter validation' do
 
     it 'adds parsed query parameters to env ' do
       get '/search', params
-      expect(last_request.env[OpenapiFirst::QUERY_PARAMS]).to eq expected_params
+      expect(last_request.env[OpenapiFirst::REQUEST].params).to eq expected_params
     end
 
     it 'skips parameter validation if no parameters are defined' do
@@ -114,7 +114,7 @@ RSpec.describe 'Query Parameter validation' do
       get '/search', params.merge(foo: 'bar')
 
       expect(last_response.status).to eq 200
-      expect(last_request.env[OpenapiFirst::QUERY_PARAMS]).to eq expected_params
+      expect(last_request.env[OpenapiFirst::REQUEST].params).to eq expected_params
     end
 
     describe 'with array query parameters' do
@@ -128,7 +128,7 @@ RSpec.describe 'Query Parameter validation' do
           }
           get '/default-style', params
           expect(last_response.status).to eq(200), last_response.body
-          parsed_parameters = last_request.env[OpenapiFirst::QUERY_PARAMS]
+          parsed_parameters = last_request.env[OpenapiFirst::REQUEST].params
           expect(parsed_parameters['strings']).to eq %w[a b c]
           expect(parsed_parameters['integers']).to eq [2, 3, 4]
         end
@@ -136,14 +136,14 @@ RSpec.describe 'Query Parameter validation' do
         it 'parses nested array' do
           get '/default-style?nested[integers]=2,3,4'
           expect(last_response.status).to eq(200), last_response.body
-          parsed_parameters = last_request.env[OpenapiFirst::QUERY_PARAMS]
+          parsed_parameters = last_request.env[OpenapiFirst::REQUEST].params
           expect(parsed_parameters['nested[integers]']).to eq [2, 3, 4]
         end
 
         it 'ignores empty query params' do
           get '/default-style'
           expect(last_response.status).to eq(200), last_response.body
-          parsed_parameters = last_request.env[OpenapiFirst::QUERY_PARAMS]
+          parsed_parameters = last_request.env[OpenapiFirst::REQUEST].params
           expect(parsed_parameters['strings']).to be_nil
         end
 
@@ -173,7 +173,7 @@ RSpec.describe 'Query Parameter validation' do
         params = {}
         get '/with-default-query-param', params
         expect(last_response.status).to eq(200)
-        parsed_parameters = last_request.env[OpenapiFirst::QUERY_PARAMS]
+        parsed_parameters = last_request.env[OpenapiFirst::REQUEST].params
         expect(parsed_parameters['has_default']).to eq true
       end
 
@@ -189,7 +189,7 @@ RSpec.describe 'Query Parameter validation' do
         params = { has_default: false }
         get '/with-default-query-param', params
         expect(last_response.status).to eq(200)
-        parsed_parameters = last_request.env[OpenapiFirst::QUERY_PARAMS]
+        parsed_parameters = last_request.env[OpenapiFirst::REQUEST].params
         expect(parsed_parameters['has_default']).to eq false
       end
     end
@@ -275,7 +275,7 @@ RSpec.describe 'Query Parameter validation' do
 
     describe 'type conversion' do
       def last_params
-        last_request.env[OpenapiFirst::QUERY_PARAMS]
+        last_request.env[OpenapiFirst::REQUEST].params
       end
 
       it 'converts to integer' do