+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add OpenapiFirst.configure
 - Add OpenapiFirst.register, OpenapiFirst.plugin
+- Replace `env[OpenapiFirst::REQUEST_BODY]`, `env[OpenapiFirst::PARAMS]` with `env[OpenapiFirst::REQUEST].body`, `env[OpenapiFirst::REQUEST].params`
 
 ## 1.0.0.beta6